requirements.txt.txt
asgiref==3.6.0
attrs==22.2.0
click==8.1.3
coverage==7.0.5
crispy-bootstrap5==0.7
Django==4.1.5
django-ckeditor==6.5.1
django-countries==7.5
django-crispy-forms==1.14.0
django-etc==1.4.0
django-filter==22.1
django-hitcount==1.3.5
django-js-asset==2.0.0
exceptiongroup==1.1.0
iniconfig==2.0.0
joblib==1.2.0
nltk==3.8.1
packaging==23.0
Pillow==9.4.0
pluggy==1.0.0
pprintpp==0.4.0
pycountry==22.3.5
pycountry-convert==0.7.2
pytest==7.2.1
pytest-cov==4.0.0
pytest-mock==3.10.0
regex==2022.10.31
repoze.lru==0.7
sqlparse==0.4.3
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.4.0